+++ commands/post_meme.js
@@ -1,6 +1,8 @@
 const config = require('../config.json');
 module.exports = {
-    post_scheduled_meme : function(client, embed, subreddit) {
+    name: 'post top meme',
+    description: 'posts',
+    execute(client, embed, subreddit) {
         const request = require('request');
         const fs = require('fs');
         const path = require('path');
@@ -55,32 +57,5 @@ module.exports = {
                 }
             }
         });
-    },
-
-    post_random_meme : function(channel, embed) {
-        try {
-            // this function will be used to post a random meme from the database
-            const fs = require('fs');
-            const path = require('path');
-
-            // files can be found in the directory 'memeData'
-            const files = ['Animemes.txt', 'DnDMemes.txt', 'Memes.txt', 'TrippinThroughTime.txt', 'DankMemes.txt', 'LotRMemes.txt', 'PrequelMemes.txt'];
-            const file_selection = files[Math.floor(Math.random() * files.length)];
-            const memes = fs.readFileSync(path.resolve(__dirname, '../memeData', file_selection)).toString().split('\n');
-            let meme = '';
-            while(meme == '') {
-                meme = memes[Math.floor(Math.random() * memes.length)];
-            }
-
-            // format the post
-            embed.setTitle('Random meme from ' + file_selection.slice(0, -4));
-            embed.setColor(0xFF0000);
-            embed.setImage(meme);
-            channel.send(embed);
-            console.log('Posted a random meme, ' + meme + ', from ' + file_selection);
-        }
-        catch(e) {
-            console.log('Error posting the random meme:', e.stack);
-        }
     },
 };
